code cleaning: move AssignedEntityGenerator out of Performer
